Refactor lib files
