+++ app/lib/constants.ts
@@ -2,4 +2,5 @@ export const CANCEL_PICKER_MESSAGES = [
   'user canceled the document picker',
   'User canceled document picker'
 ];
+
 export const DATE_FORMAT = 'MMM D, YYYY';
